// ecis-nodejs-zgwx-feat_240603/service/account-zgxw/ecissdk/model/data.ts
export interface CacheSetResp {
  result: string
  data: { data: any }
  msg?: string
}
export interface CacheGetResp {
  result: string
  data: { data: any }
  msg?: string
}

export interface CacheDelResp {
  result: string
  data: { data: any }
  msg?: string
}

export interface CacheHsetResp {
  result: string
  data: { data: any }
  msg?: string
}

export interface CacheHgetResp {
  result: string
  data: { data: any }
  msg?: string
}

export interface CacheHdelResp {
  result: string
  data: { data: any }
  msg?: string
}

export interface CacheHgetAllResp {
  result: string
  data: { data: any }
  msg?: string
}